api/requirements.txt
flask
pika
PyJWT
mysql-connector-python